requirements.txt.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic[email]==2.5.3
python-multipart==0.0.6
pytz==2024.1
celery==5.3.6
redis==5.0.1
alembic==1.13.1
python-dotenv==1.0.0